better organization, better upgrade path (versionned wrapper)
