+++ crates/rnote-cli/src/main.rs
@@ -69,7 +69,8 @@ fn main() -> anyhow::Result<()> {
                 .read_to_end(&mut bytes)?;
 
             let es = rnoteformat::load_from_bytes(&bytes)?;
-            let new_bytes = rnoteformat::save_to_bytes(es)?;
+            let new_bytes =
+                rnoteformat::save_to_bytes(es, rnoteformat::CompressionMethod::default())?;
 
             std::fs::OpenOptions::new()
                 .write(true)
@@ -174,7 +175,7 @@ fn main() -> anyhow::Result<()> {
             .map(|_| {
                 let es = engine_snapshot.clone();
                 let start = Instant::now();
-                rnoteformat::save_to_bytes(es).unwrap();
+                rnoteformat::save_to_bytes(es, rnoteformat::CompressionMethod::default()).unwrap();
                 Instant::now().duration_since(start)
             })
             .sum::<Duration>()
